telegram_worker/requirements.txt
kafka-python==2.0.2
python-telegram-bot==13.7